# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LinkedIn GPT API",
    "description": "API for accessing LinkedIn data, including top voices, search, profiles, and authentication.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "https://db828f2e-5d46-40d2-9a43-3f782fc3f1d8-00-2nhlwjy3wyleq.sisko.replit.dev/api/v1"
    }
  ],
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "description": "Search LinkedIn posts by keywords (GET method for GPT compatibility)",
        "operationId": "directSearch",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Time period for search (past-24h, past-week, past-month, day, week, month)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "past-24h"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "subscriptionId",
            "in": "query",
            "description": "Subscription ID for premium features",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Search started"
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Subscription required"
          }
        },
        "tags": ["Search"]
      }
    },
    "/api/v1/search/keywords": {
      "post": {
        "description": "Search LinkedIn posts by keywords",
        "operationId": "searchByKeywords",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keywords": {
                    "type": "string",
                    "description": "Keywords to search for"
                  },
                  "timeframe": {
                    "type": "string",
                    "description": "Time period for search (past-24h, past-week, past-month)",
                    "default": "past-24h"
                  },
                  "subscriptionId": {
                    "type": "string",
                    "description": "Subscription ID for premium features"
                  }
                },
                "required": ["keywords", "subscriptionId"]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Search started"
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Subscription required"
          }
        },
        "tags": ["Search"]
      }
    },
    "/api/v1/search/results/{searchId}": {
      "get": {
        "description": "Get results of a previous search",
        "operationId": "getSearchResults",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "description": "Search ID from a previous search request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          },
          "400": {
            "description": "Invalid request or search not completed"
          },
          "404": {
            "description": "Search not found or expired"
          }
        },
        "tags": ["Search"]
      }
    },
    "/api/v1/search/recent": {
      "get": {
        "description": "Get recent searches",
        "operationId": "getRecentSearches",
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "query",
            "description": "Subscription ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent searches"
          },
          "403": {
            "description": "Subscription required"
          }
        },
        "tags": ["Search"]
      }
    },
    "/api/v1/search/{searchId}": {
      "delete": {
        "description": "Delete a search",
        "operationId": "deleteSearch",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "description": "Search ID to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subscriptionId",
            "in": "query",
            "description": "Subscription ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search deleted successfully"
          },
          "403": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Search not found"
          }
        },
        "tags": ["Search"]
      }
    },
...
  },
...
}